layout/base/tests/bug977003_inner_5.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=977003
https://bugzilla.mozilla.org/show_bug.cgi?id=1073563
https://bugzilla.mozilla.org/show_bug.cgi?id=1094913
https://bugzilla.mozilla.org/show_bug.cgi?id=1098139
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bugs 977003, 1073563, 1094913, 1098139</title>
  <meta name="author" content="Maksim Lebedev" />
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <style>
    #target, #listener { background: yellow; }
  </style>
  <script type="application/javascript">
    var target = undefined;
    var listener = undefined;
    var test_down_target = false;
    var test_got_listener = false;
    var test_got_type = "";
    var test_got_primary = false;
    var test_lost_listener = false;
    var test_lost_type = "";
    var test_move_listener = false;
    var test_over_listener = false;
    var test_listener = false;
    var test_lost_primary = false;

    function TargetDownHandler(event) {
      logger("Target receive event: " + event.type);
      logger("Send setPointerCapture to listener");
      listener.setPointerCapture(event.pointerId);
      logger("setPointerCapture was executed");
      test_down_target = true;
    }
    function ListenerGotPCHandler(event) {
      logger("Receive event on Listener: " + event.type + "(" + event.pointerType + ")");
      listener.releasePointerCapture(event.pointerId);
      test_got_listener = true;
      test_got_type = event.pointerType;
      test_got_primary = event.isPrimary;
    }
    function ListenerLostPCHandler(event) {
      logger("Receive event on Listener: " + event.type + "(" + event.pointerType + ")");
      test_lost_listener = true;
      test_lost_type = event.pointerType;
      test_lost_primary = event.isPrimary;
    }
    function ListenerMoveHandler(event) {
      logger("Receive event on Listener: " + event.type);
      test_move_listener = true;
    }
    function ListenerOverHandler(event) {
      logger("Receive event on Listener: " + event.type);
      test_over_listener = true;
    }
    function ListenerHandler(event) {
      logger("Receive event on Listener: " + event.type);
      test_listener = true;
    }
    function logger(message) {
      console.log(message);
      var log = document.getElementById('log');
      log.innerHTML = message + "<br>" + log.innerHTML;
    }

    function prepareTest() {
      parent.turnOnPointerEvents(executeTest);
    }
    function executeTest()
    {
      logger("executeTest");
      target = document.getElementById("target");
      listener = document.getElementById("listener");
      target.addEventListener("pointerdown",          TargetDownHandler);
      listener.addEventListener("gotpointercapture",  ListenerGotPCHandler);
      listener.addEventListener("lostpointercapture", ListenerLostPCHandler);
      listener.addEventListener("pointerover",        ListenerOverHandler);
      listener.addEventListener("pointermove",        ListenerMoveHandler);
      listener.addEventListener("pointerup",          ListenerHandler);
      listener.addEventListener("pointerout",         ListenerHandler);
      var rect = target.getBoundingClientRect();
      synthesizeMouse(target, rect.width/2, rect.height/2, {type: "mousedown", isPrimary: true, inputSource: SpecialPowers.Ci.nsIDOMMouseEvent.MOZ_SOURCE_TOUCH});
      synthesizeMouse(target, rect.width/2, rect.height/2, {type: "mousemove", isPrimary: true, inputSource: SpecialPowers.Ci.nsIDOMMouseEvent.MOZ_SOURCE_TOUCH});
      synthesizeMouse(target, rect.width/2, rect.height/2, {type: "mouseup",   isPrimary: true, inputSource: SpecialPowers.Ci.nsIDOMMouseEvent.MOZ_SOURCE_TOUCH});
      finishTest();
    }
    function finishTest() {
      parent.is(test_down_target,     true, "Part 5: pointerdown event should be received by target");
      parent.is(test_got_listener,    true, "Part 5: listener should receive gotpointercapture event");
      parent.is(test_got_type,     "touch", "Part 5: gotpointercapture event should have pointerType touch");
      parent.is(test_got_primary,     true, "Part 5: gotpointercapture event should have isPrimary as true");
      parent.is(test_lost_listener,   true, "Part 5: listener should receive lostpointercapture event");
      parent.is(test_lost_type,    "touch", "Part 5: lostpointercapture event should have pointerType touch");
      parent.is(test_lost_primary,    true, "Part 5: lostpointercapture event should have isPrimary as true");
      parent.is(test_move_listener,   true, "Part 5: gotpointercapture should be triggered by pointermove");
      parent.is(test_over_listener,   true, "Part 5: listener should receive pointerover when capturing pointer");
      parent.is(test_listener,       false, "Part 5: listener should not receive any other events");
      logger("finishTest");
      parent.finishTest();
    }
  </script>
</head>
<body onload="prepareTest()">
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=977003">Mozilla Bug 977003 Test 5</a>
  <br><a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1073563">Mozilla Bug 1073563</a>
  <br><a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1094913">Mozilla Bug 1094913</a>
  <br><a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1098139">Mozilla Bug 1098139</a>
  <p id="display"></p>
  <div id="content" style="display: none">
  </div>
  <div id="listener">div id=listener</div>
  <div id="target">div id=target</div>
  <pre id="log">
  </pre>
</body>
</html>
